pyproject.toml
[tool.poetry]
name = "models"
version = "0.1.0"
description = ""
authors = ["Arkadii Paktusov <paktusov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
